Moce Gif Commands into their own subfolder
