+++ src/commands/gifs/amongus.ts
