+++ src/commands/gifs/anime.ts
