+++ src/commands/gifs/cat.ts
