+++ src/commands/gifs/pat.ts
