+++ src/commands/gifs/slap.ts
