+++ src/commands/gifs/triggered.ts
